dmnt: refactor to use sts:: namespace.
