+++ stratosphere/dmnt/source/dmnt_main.cpp
@@ -24,9 +24,7 @@
 #include <stratosphere.hpp>
 
 #include "dmnt_service.hpp"
-#include "dmnt_cheat_service.hpp"
-#include "dmnt_cheat_manager.hpp"
-#include "dmnt_config.hpp"
+#include "cheat/dmnt_cheat_service.hpp"
 
 extern "C" {
     extern u32 __start__;
@@ -96,14 +94,6 @@ void __appExit(void) {
 
 int main(int argc, char **argv)
 {
-    consoleDebugInit(debugDevice_SVC);
-
-    /* Initialize configuration manager. */
-    DmntConfigManager::RefreshConfiguration();
-
-    /* Start cheat manager. */
-    DmntCheatManager::InitializeCheatManager();
-
     /* Nintendo uses four threads. Add a fifth for our cheat service. */
     static auto s_server_manager = WaitableManager(5);
 
@@ -111,9 +101,7 @@ int main(int argc, char **argv)
 
     /* TODO: Implement rest of dmnt:- in ams.tma development branch. */
     /* server_manager->AddWaitable(new ServiceServer<DebugMonitorService>("dmnt:-", 4)); */
-
-
-    s_server_manager.AddWaitable(new ServiceServer<DmntCheatService>("dmnt:cht", 1));
+    s_server_manager.AddWaitable(new ServiceServer<sts::dmnt::cheat::CheatService>("dmnt:cht", 1));
 
     /* Loop forever, servicing our services. */
     s_server_manager.Process();
