+++ stratosphere/libstratosphere/source/pm/pm_ams.c
@@ -94,3 +94,94 @@ Result pminfoAtmosphereHasLaunchedTitle(bool *out, u64 tid) {
 
     return rc;
 }
+
+Result pmdmntAtmosphereGetProcessInfo(Handle* out, u64 *tid_out, u8 *sid_out, u64 pid) {
+    IpcCommand c;
+    ipcInitialize(&c);
+    Service *s = pmdmntGetServiceSession();
+
+    struct {
+        u64 magic;
+        u64 cmd_id;
+        u64 pid;
+    } *raw;
+
+    raw = serviceIpcPrepareHeader(s, &c, sizeof(*raw));
+
+    raw->magic = SFCI_MAGIC;
+    raw->cmd_id = 65000;
+    raw->pid = pid;
+
+    Result rc = serviceIpcDispatch(s);
+
+    if (R_SUCCEEDED(rc)) {
+        IpcParsedCommand r;
+        struct {
+            u64 magic;
+            u64 result;
+            u64 title_id;
+            FsStorageId storage_id;
+        } *resp;
+
+        serviceIpcParse(s, &r, sizeof(*resp));
+        resp = r.Raw;
+
+        rc = resp->result;
+
+        if (R_SUCCEEDED(rc)) {
+            if (out) {
+               *out = r.Handles[0];
+            } else {
+                svcCloseHandle(r.Handles[0]);
+            }
+            if (tid_out) *tid_out = resp->title_id;
+            if (sid_out) *sid_out = resp->storage_id;
+        }
+    }
+
+    return rc;
+}
+
+Result pmdmntAtmosphereGetCurrentLimitInfo(u64 *out_cur, u64 *out_lim, u32 group, u32 resource) {
+    IpcCommand c;
+    ipcInitialize(&c);
+    Service *s = pmdmntGetServiceSession();
+
+    struct {
+        u64 magic;
+        u64 cmd_id;
+        u32 group;
+        u32 resource;
+    } *raw;
+
+    raw = serviceIpcPrepareHeader(s, &c, sizeof(*raw));
+
+    raw->magic = SFCI_MAGIC;
+    raw->cmd_id = 65001;
+    raw->group = group;
+    raw->resource = resource;
+
+    Result rc = serviceIpcDispatch(s);
+
+    if (R_SUCCEEDED(rc)) {
+        IpcParsedCommand r;
+        struct {
+            u64 magic;
+            u64 result;
+            u64 cur_value;
+            u64 lim_value;
+        } *resp;
+
+        serviceIpcParse(s, &r, sizeof(*resp));
+        resp = r.Raw;
+
+        rc = resp->result;
+
+        if (R_SUCCEEDED(rc)) {
+            if (out_cur) *out_cur = resp->cur_value;
+            if (out_lim) *out_lim = resp->lim_value;
+        }
+    }
+
+    return rc;
+}
