+++ stratosphere/libstratosphere/source/pm/pm_ams.h
@@ -14,6 +14,9 @@ extern "C" {
 Result pminfoAtmosphereGetProcessId(u64 *out_pid, u64 tid);
 Result pminfoAtmosphereHasLaunchedTitle(bool *out, u64 tid);
 
+Result pmdmntAtmosphereGetProcessInfo(Handle *out, u64 *tid_out, u8 *sid_out, u64 pid);
+Result pmdmntAtmosphereGetCurrentLimitInfo(u64 *out_cur, u64 *out_lim, u32 group, u32 resource);
+
 #ifdef __cplusplus
 }
 #endif
